requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
python-dotenv
openai
PyPDF2
python-docx
Werkzeug
gunicorn
